+++ modules/dataplex/variables.tf
@@ -1,6 +1,18 @@
-variable "project_id" {
+variable "pubsub_project_id" {
   type        = string
-  description = "GCP project ID where resources will be created"
+  description = "GCP project ID where Pub/Sub resources will be created"
+}
+
+variable "monitored_folder_ids" {
+  type        = list(string)
+  description = "List of GCP folder IDs for folder-level log sinks (optional, for folder mode)"
+  default     = []
+}
+
+variable "monitored_project_ids" {
+  type        = list(string)
+  description = "List of GCP project IDs to monitor (for project mode or hybrid mode)"
+  default     = []
 }
 
 variable "masthead_service_accounts" {
@@ -10,16 +22,16 @@ variable "masthead_service_accounts" {
   description = "Masthead service account emails"
 }
 
-variable "enable_apis" {
-  type        = bool
-  description = "Enable required Google Cloud APIs"
-  default     = true
-}
-
 variable "enable_datascan_editing" {
   type        = bool
   description = "Enable permissions for creating and editing Dataplex DataScans"
   default     = false
+}
+
+variable "enable_apis" {
+  type        = bool
+  description = "Enable required Google Cloud APIs"
+  default     = true
 }
 
 variable "labels" {